feat(app): inject PostgresWorkoutStore into WorkoutHandler

+++ internal/store/workout_store.go
@@ -32,7 +32,7 @@ func NewPostgresWorkoutStore(db *sql.DB) *PostgresWorkoutStore {
 
 type WorkoutStore interface {
 	CreateWorkout(*Workout) (*Workout, error)
-	GetWorkoutByID(id int) (*Workout, error)
+	GetWorkoutByID(id int64) (*Workout, error)
 }
 
 func (pg *PostgresWorkoutStore) CreateWorkout(workout *Workout) (*Workout, error) {
